tests/wpt/web-platform-tests/websockets/opening-handshake/002.html

<!doctype html>
<title>WebSockets: valid handshake</title>
<meta name=timeout content=long>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js?pipe=sub></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/echo');
  ws.onopen = t.step_func(function(e) {
    ws.onclose = t.step_func(function(e) {
      assert_equals(e.wasClean, true);
      ws.onclose = t.step_func(function() {assert_unreached()});
      setTimeout(t.step_func(function() {t.done();}), 50)
    })
    ws.close();
  });
  ws.onerror = ws.onmessage = ws.onclose = t.step_func(function() {assert_unreached()});
}, null, {timeout:9900});
</script>
